+++ guillotina_elasticsearch/__init__.py
@@ -6,12 +6,6 @@ from guillotina.content import IResourceFactory
 from guillotina.utils import get_dotted_name
 
 
-# import os
-
-
-ELASTIC6 = False  # os.environ.get("ES_VERSION", "").strip() == "6"
-
-
 def default_refresh():
     return False