+++ guillotina_elasticsearch/interfaces.py
@@ -1,15 +1,12 @@
 from guillotina.async_util import IAsyncUtility
 from guillotina.catalog.types import BasicParsedQueryInfo
 from guillotina.interfaces import ICatalogUtility
-from guillotina_elasticsearch import ELASTIC6
 from zope.interface import Interface
 
 import typing
 
 
 DOC_TYPE = "_doc"
-if ELASTIC6:
-    DOC_TYPE = "doc"
 
 
 SUB_INDEX_SEPERATOR = "__"