+++ guillotina_elasticsearch/migration.py
@@ -21,7 +21,6 @@ from guillotina.utils import get_content_path
 from guillotina.utils import get_current_container
 from guillotina.utils import get_current_transaction
 from guillotina.utils import get_security_policy
-from guillotina_elasticsearch import ELASTIC6
 from guillotina_elasticsearch.events import IndexProgress
 from guillotina_elasticsearch.interfaces import DOC_TYPE
 from guillotina_elasticsearch.interfaces import IIndexActive
@@ -288,12 +287,7 @@ class Migrator:
         Missing ones are ignored and we don't care about it.
         """
         next_mappings = await self.conn.indices.get_mapping(self.work_index_name)
-        next_mappings = next_mappings[self.work_index_name]["mappings"]
-
-        if ELASTIC6:
-            next_mappings = next_mappings[DOC_TYPE]["properties"]
-        else:
-            next_mappings = next_mappings["properties"]
+        next_mappings = next_mappings[self.work_index_name]["mappings"]["properties"]
 
         existing_index_name = await self.index_manager.get_real_index_name()
         try:
@@ -302,12 +296,9 @@ class Migrator:
             # allows us to upgrade when no index is present yet
             return next_mappings
 
-        existing_mappings = existing_mappings[existing_index_name]["mappings"]
-
-        if ELASTIC6:
-            existing_mappings = existing_mappings[DOC_TYPE]["properties"]
-        else:
-            existing_mappings = existing_mappings["properties"]
+        existing_mappings = existing_mappings[existing_index_name]["mappings"][
+            "properties"
+        ]
 
         new_definitions = {}
         for field_name, definition in next_mappings.items():